+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project mostly adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.0] - 2020-05-09
+
+## Added
+- Support for Tailwind 1.3 and 1.4 features
+  - Added `purgecss` and `target` options
+  - Set the variants for the new color opacity utilities
+  - Disabled the new `space` and `divide` utilities by default
+
 ## [3.0.2] - 2020-04-19
 
 ### Changed
@@ -390,7 +398,8 @@ and this project mostly adheres to [Semantic Versioning](https://semver.org/spec
 
 Initial release
 
-[Unreleased]: https://github.com/benface/tailwindcss-config/compare/v3.0.2...HEAD
+[Unreleased]: https://github.com/benface/tailwindcss-config/compare/v3.1.0...HEAD
+[3.1.0]: https://github.com/benface/tailwindcss-config/compare/v3.0.2...v3.1.0
 [3.0.2]: https://github.com/benface/tailwindcss-config/compare/v3.0.1...v3.0.2
 [3.0.1]: https://github.com/benface/tailwindcss-config/compare/v3.0.0...v3.0.1
 [3.0.0]: https://github.com/benface/tailwindcss-config/compare/v3.0.0-beta.3...v3.0.0